layer_app/QIBTrader_Pro/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0.0
zeroc-ice>=3.7.0